requirements.forwarder.txt
fastapi
uvicorn[standard]
httpx>=0.27.0
pydantic>=2.12.3
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-logging
